Disambiguate buffer names
We keep track of buffer `head` and `tail`, which is basically the
directory and basename. We start off with just the `tail` as the buffer
`name`. If there are multiple occurrences of the same `name`, we prepend
`tail` with the last element from `head`. We keep using more and more
elements from `head` until all buffer names are unique. New buffers are
not included in the disambiguate process.